Extract Joy UI config to a separate directory
